Fix subject creation (#90)
Close #90
